converted the usage of a few long and longlong to int32_t and int64_t
git-svn-id: http://euclid.esac.esa.int/svn/EC/SGS/SDC/CH/Projects/Elements/trunk@2872 463500fe-d6be-4184-9b20-47717e75ecac

+++ ElementsKernel/ElementsKernel/Sleep.h
@@ -1,6 +1,7 @@
 #ifndef ELEMENTSKERNEL_SLEEP_H_
 #define ELEMENTSKERNEL_SLEEP_H_
 
+#include <cstdint>
 #include "ElementsKernel/Kernel.h"
 
 namespace Elements {
@@ -11,7 +12,7 @@ ELEMENTS_API void normalSleep(int sec);
 
 /// Small variation on the sleep function for nanoseconds sleep.
 /// @author Marco Clemencic
-ELEMENTS_API void nanoSleep(long long nsec);
+ELEMENTS_API void nanoSleep(int64_t nsec);
 
 }
